Fixes #8625: Split linters to run for a specific file extension (#8656)
* Splited linter by file extension types

* Run linter using semaphore method

* Used Semaphore method for linting

* Fix yarn.lock

* Added more isolation

* Fix yarn.lock

* Added manager for third party lint checks

* Fix typo

* Splitted the linter into sub linters

* Moved under linter directory

* Addressed review changes

* Changed pre_commit_linter path

* Temporarily disabled BlankLineBelowFileOverviewChecker

* Fixed lint error

* Added test for redirect_stdout

* Removed html directive check

* Addressed review comments

* Used list.extend

* Fixed lint errors

* Addressed review comments

* Increased concurrency limit to 2

* Addressed review comments

* Fixed lint errors

* Semaphore fixed

* Addressed review comments

* Addressed review comments.

* Addressed review comments

* Addressed review comments

* Addressed review comments
