+++ .github/workflows/lint.yml
@@ -18,4 +18,4 @@ jobs:
       - name: Third Party Size Check
         run: python -m scripts.third_party_size_check
       - name: Run Lint Checks
-        run: python -m scripts.pre_commit_linter --path=. --verbose
+        run: python -m scripts.linters.pre_commit_linter --path=. --verbose
