+++ core/tests/protractor_desktop/embedding.js
@@ -259,9 +259,9 @@ describe('Embedding', function() {
       waitFor.pageToFullyLoad();
 
       // TODO(#7622): Remove this sleep. Remove the todo in
-      // scripts/pre_commit_linter.py pointing to the same issue This sleep was
-      // placed here as a temporary measure because the wait in the line above
-      // does not wait for the iframe to load fully.
+      // scripts/linters/pre_commit_linter.py pointing to the same issue This
+      // sleep was placed here as a temporary measure because the wait in the
+      // line above does not wait for the iframe to load fully.
       browser.sleep(6000);
 
       expect(driver.findElement(by.css('.protractor-test-float-form-input'))
