+++ scripts/linters/linter_utils_test.py
@@ -20,6 +20,7 @@ from __future__ import absolute_import  # pylint: disable=import-only-modules
 from __future__ import unicode_literals  # pylint: disable=import-only-modules
 
 import collections
+import tempfile
 
 from core.tests import test_utils
 import python_utils
@@ -170,3 +171,19 @@ class MemoizeTest(test_utils.GenericTestBase):
         count_calls(5, 6, d=3, c=2)
         self.assertEqual(call_counter[(5, 6, 3, 2)], 0)
         self.assertEqual(call_counter[(5, 6, 2, 3)], 1)
+
+
+class RedirectStoutTest(test_utils.GenericTestBase):
+    """Tests for the redirect_stdout function."""
+
+    def test_redirect_stdout(self):
+        temp_file = tempfile.NamedTemporaryFile()
+
+        with python_utils.open_file(temp_file.name, 'r+') as temp_file_contents:
+            with linter_utils.redirect_stdout(temp_file_contents):
+                python_utils.PRINT('This is a test')
+            temp_file_contents.seek(0)
+            data = temp_file_contents.read()
+        temp_file.close()
+
+        self.assertEqual(data, 'This is a test\n')
