+++ scripts/run_backend_tests.py
@@ -44,7 +44,6 @@ from __future__ import absolute_import  # pylint: disable=import-only-modules
 from __future__ import unicode_literals  # pylint: disable=import-only-modules
 
 import argparse
-import datetime
 import importlib
 import inspect
 import multiprocessing
@@ -58,6 +57,7 @@ import time
 import python_utils
 
 from . import common
+from . import concurrent_task_utils
 from . import setup
 from . import setup_gae
 
@@ -99,13 +99,9 @@ COVERAGE_MODULE_PATH = os.path.join(
     'coverage-%s' % common.COVERAGE_VERSION, 'coverage')
 
 TEST_RUNNER_PATH = os.path.join(os.getcwd(), 'core', 'tests', 'gae_suite.py')
-LOG_LOCK = threading.Lock()
-ALL_ERRORS = []
 # This should be the same as core.test_utils.LOG_LINE_PREFIX.
 LOG_LINE_PREFIX = 'LOG_INFO_TEST: '
 _LOAD_TESTS_DIR = os.path.join(os.getcwd(), 'core', 'tests', 'load_tests')
-
-MAX_CONCURRENT_RUNS = 25
 
 _PARSER = argparse.ArgumentParser(description="""
 Run this script from the oppia root folder:
@@ -137,19 +133,6 @@ _PARSER.add_argument(
     action='store_true')
 
 
-def log(message, show_time=False):
-    """Logs a message to the terminal.
-
-    If show_time is True, prefixes the message with the current time.
-    """
-    with LOG_LOCK:
-        if show_time:
-            python_utils.PRINT(
-                datetime.datetime.utcnow().strftime('%H:%M:%S'), message)
-        else:
-            python_utils.PRINT(message)
-
-
 def run_shell_cmd(exe, stdout=subprocess.PIPE, stderr=subprocess.PIPE):
     """Runs a shell command and captures the stdout and stderr output.
 
@@ -164,11 +147,12 @@ def run_shell_cmd(exe, stdout=subprocess.PIPE, stderr=subprocess.PIPE):
     last_stdout = last_stdout_str.split('\n')
 
     if LOG_LINE_PREFIX in last_stdout_str:
-        log('')
+        concurrent_task_utils.log('')
         for line in last_stdout:
             if line.startswith(LOG_LINE_PREFIX):
-                log('INFO: %s' % line[len(LOG_LINE_PREFIX):])
-        log('')
+                concurrent_task_utils.log(
+                    'INFO: %s' % line[len(LOG_LINE_PREFIX):])
+        concurrent_task_utils.log('')
 
     result = '%s%s' % (last_stdout_str, last_stderr_str)
 
@@ -176,39 +160,6 @@ def run_shell_cmd(exe, stdout=subprocess.PIPE, stderr=subprocess.PIPE):
         raise Exception('Error %s\n%s' % (p.returncode, result))
 
     return result
-
-
-class TaskThread(threading.Thread):
-    """Runs a task in its own thread."""
-
-    def __init__(self, func, verbose, semaphore, name=None):
-        super(TaskThread, self).__init__()
-        self.func = func
-        self.output = None
-        self.exception = None
-        self.verbose = verbose
-        self.name = name
-        self.semaphore = semaphore
-        self.finished = False
-
-    def run(self):
-        try:
-            self.output = self.func()
-            if self.verbose:
-                log('LOG %s:' % self.name, show_time=True)
-                log(self.output)
-                log('----------------------------------------')
-            log('FINISHED %s: %.1f secs' %
-                (self.name, time.time() - self.start_time), show_time=True)
-        except Exception as e:
-            self.exception = e
-            if 'KeyboardInterrupt' not in python_utils.convert_to_bytes(
-                    self.exception.args[0]):
-                log('ERROR %s: %.1f secs' %
-                    (self.name, time.time() - self.start_time), show_time=True)
-        finally:
-            self.semaphore.release()
-            self.finished = True
 
 
 class TestingTaskSpec(python_utils.OBJECT):
@@ -229,59 +180,6 @@ class TestingTaskSpec(python_utils.OBJECT):
             exc_list = [sys.executable, TEST_RUNNER_PATH, test_target_flag]
 
         return run_shell_cmd(exc_list)
-
-
-def _check_all_tasks(tasks):
-    """Checks the results of all tasks."""
-    running_tasks_data = []
-
-    for task in tasks:
-        if task.isAlive():
-            running_tasks_data.append('  %s (started %s)' % (
-                task.name,
-                time.strftime('%H:%M:%S', time.localtime(task.start_time))
-            ))
-
-        if task.exception:
-            ALL_ERRORS.append(task.exception)
-
-    if running_tasks_data:
-        log('----------------------------------------')
-        log('Tasks still running:')
-        for task_details in running_tasks_data:
-            log(task_details)
-
-
-def _execute_tasks(tasks, semaphore):
-    """Starts all tasks and checks the results.
-    Runs no more than the allowable limit defined in the semaphore.
-
-    Args:
-        tasks: list(TestingTaskSpec). The tasks to run.
-        semaphore: threading.Semaphore. The object that controls how many tasks
-            can run at any time.
-    """
-    remaining_tasks = [] + tasks
-    currently_running_tasks = []
-
-    while remaining_tasks:
-        task = remaining_tasks.pop()
-        semaphore.acquire()
-        task.start()
-        task.start_time = time.time()
-        currently_running_tasks.append(task)
-
-        if len(remaining_tasks) % 5 == 0:
-            if remaining_tasks:
-                log('----------------------------------------')
-                log('Number of unstarted tasks: %s' % len(remaining_tasks))
-            _check_all_tasks(currently_running_tasks)
-        log('----------------------------------------')
-
-    for task in currently_running_tasks:
-        task.join()
-
-    _check_all_tasks(currently_running_tasks)
 
 
 def _get_all_test_targets(test_path=None, include_load_tests=True):
@@ -405,7 +303,8 @@ def main(args=None):
             include_load_tests=include_load_tests)
 
     # Prepare tasks.
-    concurrent_count = min(multiprocessing.cpu_count(), MAX_CONCURRENT_RUNS)
+    max_concurrent_runs = 25
+    concurrent_count = min(multiprocessing.cpu_count(), max_concurrent_runs)
     semaphore = threading.Semaphore(concurrent_count)
 
     task_to_taskspec = {}
@@ -413,20 +312,21 @@ def main(args=None):
     for test_target in all_test_targets:
         test = TestingTaskSpec(
             test_target, parsed_args.generate_coverage_report)
-        task = TaskThread(
+        task = concurrent_task_utils.create_task(
             test.run, parsed_args.verbose, semaphore, name=test_target)
         task_to_taskspec[task] = test
         tasks.append(task)
 
     task_execution_failed = False
     try:
-        _execute_tasks(tasks, semaphore)
+        concurrent_task_utils.execute_tasks(tasks, semaphore)
     except Exception:
         task_execution_failed = True
 
     for task in tasks:
         if task.exception:
-            log(python_utils.convert_to_bytes(task.exception.args[0]))
+            concurrent_task_utils.log(
+                python_utils.convert_to_bytes(task.exception.args[0]))
 
     python_utils.PRINT('')
     python_utils.PRINT('+------------------+')
