+++ scripts/run_presubmit_checks.py
@@ -28,9 +28,9 @@ import subprocess
 import python_utils
 
 from . import common
-from . import pre_commit_linter
 from . import run_backend_tests
 from . import run_frontend_tests
+from .linters import pre_commit_linter
 
 _PARSER = argparse.ArgumentParser(description="""
 Run this script from the oppia root folder prior to opening a PR:
